shard: pass meta problem as error, drop HasMeta() from *Res
If we have no error and got a result -- no one cares about HasMeta(). If
we've got an error then currently some codes cares about meta status, but it
can be passed as an additional error as well. Related to #3024, in reality
meta should be internal to the shard, but this change at least avoids
excessive booleans.

Signed-off-by: Roman Khimov <roman@nspcc.ru>