fstree: introduce combined file versioning
We may want to have something more complex in future versions (like #2925 or
others) and we don't have a lot of safe bytes to use in place of 0x7F. So
let's have a byte for format version.

This is an incompatible change, but this code was never in production.

Signed-off-by: Roman Khimov <roman@nspcc.ru>