fstree: fix errorHandler behavior
I believe this bug was introduced right with 7df50297cd9122242b2bff2f7f85a6604083e8f8,
this code is a straight copy-paste from bbcz, but bbcz operates in a callback,
so it must return anyway and if errorHandler returns non-nil value is stops
the iteration. In FSTree this leads to iteration stopping in any event if
errorHandler is invoked and this shouldn't be the case.

Signed-off-by: Roman Khimov <roman@nspcc.ru>