ir, morph/container: put container placement changes in a single TX
It is the most secure way to ensure changes are done in sync and either all at
once, or none of them. It is still can be a problem to update too long
placement vectors in a single TX since there is always a certain limit, but it
can be reconsidered separately.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>