node/put: add first object ID to the metadata pushed on chain
It will always have some first object ID: if it is the same as object ID
itself -- it is the first part of the chain; if it differs -- this is any other
part of the split; if it is 32 zeros (reserved "zero" object ID) -- this object
is a small one and does not belong to any object split. Closes #2935.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>