node: use UpdateState for heartbeats when we can
Most of the time we don't need a complete AddPeer/AddNode. Fixes #1914.

This also changes IR slightly, we update epoch on UpdateState now to work with
v1 nodes correctly and cleanup is completely disabled for v2 nodes since it's
performed by the contract and this code is to be deleted in future.

Signed-off-by: Roman Khimov <roman@nspcc.ru>