metabase: fix that parent is not deleted
When the parent gets deleted, the `objectSDK.SplitInfoError` error is returned
when calling get and exits the function without deleting the object from the
metabase. But for some reason, the link was deleted from the parent bucket,
although it is not there. Currently, it is the parent object that is being
deleted by the link.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>