node: fix race between heartbeat and state updater
The first heartbeat attempt after the node bootstrap can end up this way:

    warn    neofs-node/netmap.go:189        can't send heartbeat tx {"error": "incorrect current network map status OFFLINE, restart recommended"

Which is not critical since the next one will succeed, but the root of the
problem is _two_ asynchronous handlers started for a new epoch, one of them
updating local node data and the other sending heartbeats. This problem
existed even before nodeV2, but with the old heartbeat mechanism it wasn't
visible.

I doubt we need two routines here, everything can be performed with a single
one sequentially and correctly.

Signed-off-by: Roman Khimov <roman@nspcc.ru>