morph: fix error from `CalculateNonceAndVUB`
The alphabetic key change event comes from the main chain with its hash, and
when calculating vub, an error is returned because there is no hash in the FS
chain. So for designation event of the mainnet RoleManagement contract
transaction hash is set to zero.
`CalculateNonceAndVUB` can now handle zero hashes and calculate vub rounded.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>