node/metabase: Store int attributes in both formats for SearchV2
Previously, int attributes were stored in 0x01-prefixed key space
preserving both numeric and lexicographic order. These keys were
separated from plain non-int attributes stored with 0x02 prefix. The
advantage of this approach was the lower consumption of database space
for indexes. Еhe disadvantage was the inconvenience of processing
non-NUM filters like NE or PREFIX: it was necessary to jump from one
index to another with a possible loss of order.

Given that increasing the DB size with an additional index does
not cause problems yet, and iteration is accelerated in runtime and
simplified in code, this adds an integers to the plain index.

Fixes #3161. Closes #3152 (wontfix). Refs #3130. Refs #3058.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>