event: rework pool model into queues
A limited non-blocking pool handles all types of events which leads to:
 * asynchronous processing (#2214)
 * throwing away events we can't process when too many of them happen at
   the same time

What's done here is:
 * three natural upper-level goroutines for different event types
 * concurrent notary event processing with a separate fixed pool of routines
 * sequential processing of block and contract events in their own routines
 * a queue of events in channels

The good thing is that it's ordered and doesn't throw away events, it blocks.
The bad things is that it can block. But this just means we're not capable
of processing events at the speed they happen, so some reconnection should
happen at the upper layer.

I believe it fixes #2214, the only concurrency here is blocks vs notifications,
but given the way we process blocks it's almost irrelevant. We just store them
in the DB and check against timers.

Signed-off-by: Roman Khimov <roman@nspcc.ru>