object/search2: Fix disorder of some system attributes on merge
Binary and string representation of some attributes differ. When sorting
a search result, the metabase compares the binary format. When merging
responses from multiple nodes, string sorting is used. For all user
attributes, this is essentially the same. However, for some system
ones, this is not the case. For example, object IDs and owners are
base58btc strings which do not preserve byte order.

Now sorting at metabase and merging server levels are synchronized, and
disorder is fixed.

Fixes #3160. Refs #3058.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>