node/metabase: Interrupt search by primary filter faster
Primary filters are applied to sorted meta bucket's elements. This
allows us to speed up execution when we reached mismatch at some point:
 - for EQ/PREFIX/LT/LE/GE matchers, any next item will definitely fail;
 - for GT matcher, any next item will definitely fail if some previous
 element matched.

Now the meta bucket iterator breaks when the mentioned conditions are
reached. This speeds up the processing of most search queries,
especially as the number of objects increases.

Refs #3058.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>