morph: recreate fallback transactions for retries
The primary problem is:

    error client/notary.go:455 can't send notary request after retries {"retries": 3, "error": "failed to submit notary request: Invalid signature (-508) - witness #1: signature check failed: invalid signature"}

which happens because SendRequest() changes the fallback and signs it again,
but one of the parameters involved is current blockchain height which can
change while we're retrying and since the transaction hash is cached and
not changed we're signing a wrong one.

But what follows from this is that stable FB hashes were always an illusion
even though 38a5bb12655ac9e476a8480fb94ee8d5c1a18a84 tried to provide them.
They can be provided with a big more dancing, but it doesn't seem like we're
suffering in any way because of their absence.

So this takes the path of the least resistance and simplifies the code to
autocreate fallbacks when needed. Whatever they are they will work OK for us.

Signed-off-by: Roman Khimov <roman@nspcc.ru>