node/metabase: Migrate objects to metabucket in batches
Previously, any migration to newer metabase version was performed within
the single BoltDB transaction. Thus, V3->4 migration tried to overtake
all objects to the metabucket in one tx. Practice has shown that such an
approach can delay migration for hours and even days if the objects
number in the millions.

This brings two changes to the metabase init procedure:
 1. current version number is read in a separate BoltDB RO transaction;
 2. any metabase version upgrade is done in separate BoltDB transaction(s);
 3. V3->4 migration writes objects' metadata to the metabucket in
   batches of 1000 size.

2 was needed to achieve 3. Overall this is a good change as upgrading a
version is an atomic operation. 3 was tested in practice: it sped up
processing of 9KK objects from almost a week (!) to half an hour.

Closes #3202.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>