node/metabase: Ignore already objects migrated to metabucket earlier
With f105b24ae3839fa1828a7d875d62a57bfd98e790, metabase migration V3->4
splits the whole process in 1000-object batches. Each batch is done
in a single BoltDB transaction. Transaction multiplicity automatically
introduced the possibility of partial filling of the metabucket, for
example, when node was suffered accidental shutdown. Therefore, when
restarting the migration procedure, some objects are ready, and
they are not skipped. Although it is worth noting that repeating the
object write is not an error, it is just no-op in terms of storage.

This makes migrator to check whether next object has already been stored
in the metabucket, and skip it if so.

Refs #3202.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>