netmap: always approve v2 node bootstrap
SNs on old networks with v1 node netmap send two bootstrap requests one after
another, v1 and v2. But IR approves only one because of "updated" check:

    neofs-ir[1458]: info        netmap/handlers.go:45        notification        {"type": "add peer"}
    neofs-ir[1458]: info        netmap/handlers.go:64        notification        {"type": "add node"}
    neofs-ir[1458]: info        netmap/process_peers.go:66        approving network map candidate        {"key": "03aeff8a19f0202090afb0916b1c00b432321be7e8623a06c9b9b5db8ee5c053a4"}

Even though the system works with the old map fine we can't switch to v2
because that map is either empty or has 1-2 node (sometimes v2 is approved
before v1 due to concurrency).

This is a minimal fix to always approve correct v2 nodes, then after the switch
we'll simplify all of this code.

Signed-off-by: Roman Khimov <roman@nspcc.ru>