node/metabase: Fix collecting of converted attributes in SearchV2
Some object attributes are stored in the meta index and are passed
through the protocol in different ways. For example, IDs are transmitted
as Base58 strings. Previously, metabase's search method called by the
`ObjectService.SearchV2` handler did not convert these attributes. In
particular, it was found in a server-side proto encoding error: ID can
contain byte sequences that are incorrect for UTF-8. Although this is an
indirect consequence, it helped to discover the root cause.

This makes SearchV2 handler to properly convert requested primary
attribute and respond with them according to the protocol.

Fixes #3230. Refs #3058.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>