ir: do not use neofsid for container checks
neofsid is obsolete, this mechanism is completely irrelevant for proper
N3 accounts. Notice that this makes keys mandatory in signatures for now,
but that's expected to be OK (no one actually uses neofsid anyway).

Notice that container deletion requests don't have any key attached, so we
recover it from the container data for now.

Signed-off-by: Roman Khimov <roman@nspcc.ru>