node/object: Check that bearer token is signed by its issuer
Previously, a bearer token could bypass verification by being signed
with any key if issuer field was explicitly set equal to the container
owner. This was caused upgrade to NeoFS SDK with
https://github.com/nspcc-dev/neofs-sdk-go/commit/06e0977ff0243583b958e15ee76d2830ad23623b
that changed `bearer/Token.ResolveIssuer` method.

This fixes the bug and adds corresponding unit test. Code is refactored
a bit to simplify testing.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>