meta: fix container creation deadlock
It is prohibited to make RPC calls if you are not reading notification channel.
If last container removal and new container creation are events from a single
block, 7ec47bd54033766026d5c936f007c3726db36740 deadlocks in its
(re)subscription subroutines. The logic will be improved in the future when
every notification is handled in a strict order: #3259.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>