metabase: fix middle child objects being treated as root ones
They have previous and first OIDs, but they don't have any proper parent
header. It's all accounted for in object.HasParent(), but for whatever reason
we had some special handling of parents for searchv2 metadata. In fact old
putUniqueIndexes() did exactly this for root object indexing:

    // index root object
    if obj.Type() == objectSDK.TypeRegular && !obj.HasParent() {

SearchV2 metadata is not much different and we don't need any parameter for
PutMetadataForObject(), it all can be derived from the object itself.

Signed-off-by: Roman Khimov <roman@nspcc.ru>