node/object: Optimize filtered attribute-less SearchV2 queries
Previously, when `SearchV2Request` included filters but not attributes,
server iterated OID key space of the meta bucket just like for
unfiltered queries. This could lead to slow query processing on
containers with big number of objects.

This optimizes handling of mentioned queries by requesting first
filtered attribute from local storage and other nodes so that order and
continuity can be ensured when merging the results. As a result, the
behavior completely coincides with the attributed case, but the response
is simply cut off.

It should be noted that described approach is enabled for in-container
SNs only. Otherwise, request is processed same as before.

Closes #3329.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>