node/object: Skip already removed members in tombstone verification
SN checks the legitimacy of deleting objects when receiving a tombstone
for them. Tombstone can be put in parallel, so when one SN validates it,
another could already accept it and mark the members as deleted.
In this case, verifying SN receives ALREADY_REMOVED status on HEAD
request. Previously, late SN failed the verification with this status,
although it signals that tombstone succeeded remotely. This could lead
to failure of the whole tombstone PUT op.

This fixes the problem by ignoring status error from HEAD response.

Fixes #3327.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>