object/get: fix that object can't be got without a link object
This problem occurs when the link object has not yet been deleted from the
metabase by gc. However, there is no problem if it has already been deleted.
In that case it is an `object not found` error, so decided to add a fallback in
the case where metabase gives a link ID but cannot get an object and there is no
last part.

Closes #3275.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>