shard: fix expired lock obj test
We have meta here with epoch 10 and shard with epoch 5 after the event.
Expired object handler marks the object as garbage, but GC needs time to
collect it. By default it runs every 10 seconds, so 3s `require.Eventually()`
is not sufficient for object to be deleted, it's still there, marked for
deletion and expired at the same time. Now our exists() code checks for
expiration first, so if it's corrected wrt #3321 it returns "expired"
instead of "marked for GC". It shouldn't be a problem for users since the
object can't be retreived anyway, but for this test if we want to check
for GC to actually delete the object we need to adjust its timing parameters.

Signed-off-by: Roman Khimov <roman@nspcc.ru>