shard: fix expiration test
We have metabase here that thinks that the epoch is MaxUint64/2, so all objects
put into the shard are already expired for the metabase and attempt to get them
can technically lead to "expired" answer. Previously this wasn't noticeable
because of the expiration check behavior (#3321). In this particular test we're
checking object expirations by GC, so metabase shouldn't affect the result,
setting its epoch to 0 leads to the fact that objects are never expired for the
metabase, but shard-level GC still finds and marks them because it uses a
different epoch. I think that's exactly what we want to check for, so 0 is OK.

Signed-off-by: Roman Khimov <roman@nspcc.ru>