node/object: Cache results of bearer token verifications
Bearer tokens are checked on every request. Sometimes the same token is
used for a class of repeated requests, such as those sent from the same
protocol gate to the same container. Some of the checks depend on the
network state (e.g. expiration), while the result of the rest remains
unchanged with a fixed token and request context. This opens up
opportunities for caching.

This extends SN with LRU cache for up to 1000 static bearer tokens'
validation results.

Refs #3243.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>