shard: check metabase before removing things from write cache
Resync metabase test avoids putting things into write cache, but when modified
to not do so it quickly detects this problem. Inhume() operation can have zero
effect depending on the metabase state and in this case the object should be
kept while we were _always_ dropping it from write cache, so it was there
we're out of luck, it's lost.

Signed-off-by: Roman Khimov <roman@nspcc.ru>