sn/object/acl: Use cached container SN lists for ACL processing
To control object access, SN applies container storage policy to the
current and, if needed, previous network map in order to distinguish
requests sent by container nodes. Previously, SN performed this action
on each request. Under high load, a request for the same container from
the same container node can be made in parallel in large quantities. A
situation of a large flow of parallel requests on the same container may
arise. In this case, policy application which includes HRW sorting is
done again and again. If the server's CPU is not a strong point, this
leads to increased latency. Since container entry does not change
within an epoch, caching the results can optimize the process.

This makes SN IAM module to reuse cache of containers' storage policy
application results that were previously enabled for other server tasks.

Refs #1803. Refs #2692.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>