node/object: Cache results of bearer token verifications
Bearer tokens are checked on every request. Sometimes the same token is
used for a class of repeated requests, such as those sent from the same
protocol gate to the same container. In most cases, the result remains
unchanged with a fixed token and request context. This opens up
opportunities for caching.

This makes SN to cache up to 1000 LRU bearer token common check results
until the next epoch.

Refs #3243.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>