sn/object: Forward original GET/HEAD/RANGE request to container members
Previously, when proxy SN forwarded read object requests to the
container, it updated the request as follows:
 1. meta header was wrapped into the new one with decremented TTL;
 2. verification header was wrapped into to the new one with added
   signatures for original parts.

This resulted in the following costs:
 - additional calculation of two signatures on proxy SN;
 - more expensive request serialization, more data over the network;
 - more expensive request deserialization on container SN;
 - additional verification of two signatures on container SN.

At the same time, apart from the mentioned actions, container SN did not
use additional data for processing. So if a node doesn't update TTL and
add signatures, then it's just reduces the resources spent without
changing the behavior.

This makes proxy SN to not resign object GET/HEAD/RANGE requests on
forwarding.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>