metabase: rework Get() with searchV2 index
The only user of Get() left is neofs-lens and it's a debugging tool, so its
job is to return whatever data metabase has. Since we're going to drop headers
the only data left is searchV2 index which allows to retrieve a substantial
part of the header, but not all of it.

PutBinary test doesn't have any value with this change, we no longer return
binaries.

Signed-off-by: Roman Khimov <roman@nspcc.ru>