object: drop redundant EACL check for Search
Search result is just IDs, the result doesn't have any new additional data
compared to request processing time, if we had an object header we still
have it, if we had no object header we still don't have it, the request itself
hasn't change, so this check is no-op, it can't affect the result.

Notice that SearchV2 doesn't do this at all. Refs #3095.

Signed-off-by: Roman Khimov <roman@nspcc.ru>