object: don't check EACL twice when it's not needed
Get and Head handlers check EACL before doing anything (just like other
handlers do), but if there are rules based on object headers and this object
is not stored locally the action would be allowed. Once we get a header from
the other node EACL can and will always be rechecked which is suboptimal since
in most cases it's the same simple (non-filter-dependent) set of rules. The
patch tries to optimize it by avoiding rechecks if there was a rule match.
If that's the case another EACL check would yield the same (positive) result
anyway, so it's just a waste of time.

Notice that from the security perspective this is still questionable since if
the node storing object (erroneously) returns data this means the data can
technically be retrieved by asking it directly.

Closes #3095.

Signed-off-by: Roman Khimov <roman@nspcc.ru>