storage/shard: Fix metabase resync for object+tombstone+lock case
Resync metabase process of storage shard overtakes objects from BLOB
storage to the metabase. There can be object and tombstone + lock
objects associated with it. Processing order may vary during resync.

Previously, resync behavior could differ from the SN runtime. In
particular, if the tombstone is PUT first, and then the lock - all three
objects are located in the metabase. Then on resync, if the lock were
handled before the tombstone, the process was interrupted with the
'object locked' error. If, on the contrary, tombstone was ahead of
tombstone, then when tombstone written, the object marked as garbage
and then deleted, that is, the lock did not protect it.

This fixes both described scenarios and tests that resync result of such
a group of objects does not depend on the order.

Reverts 504abdda06bb6ac977461e1d1c40a0b6cec158b8 and
ebcafd33ec4404bdcf1042e7a8a2abdaf5dd2477 as no longer possible.

Closes #3436.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>