ir: drain and close deployment notification channels, fix #3237
The problem is that deploy code subscribes and never unsubscribes while it
spawns some readers that are bound to deploy context. So once Deploy() is done
they can (and should) exit, but we still haven't performed unsubscription, so
there is a race between events and unsubscription that's especially annoying
with small block times.

Ideally it's deploy code that should be doing this (it spawns receivers), but
draining here concurrently to deploy threads should be sufficient too.

Signed-off-by: Roman Khimov <roman@nspcc.ru>