workflows: don't build Dockers in PRs
The only thing it does in PR is ensures that Docker images can be built,
they're not pushed into any registry, so they can't be used and no one
needs the result. Building in master is sufficient for the purpose of
finding failed Docker builds (they don't break often), but we better
avoid doing it in PRs since it takes a considerable amount of time
blocking runners available to project/organization.

Signed-off-by: Roman Khimov <roman@nspcc.ru>