meta: get container from the same channel asynchronously, fix #3491
Meta uses WSClient of its own that can be connected to a different node with
a different state which means

    2025-07-30T07:43:02.941Z	error	meta/notifications.go:140	can't get container data	{"service": "meta data", "notification container": "616c11fbe324d23e77886ca062bc57633217082373ee0cd9ba984aeae4daafc6", "error": "read E36vNnxN7Bs16HPMf5EBWTk29aSVrDDg2Cx64nYYgbdR container: status: code = 3072 message = container not found"}

can happen easily. Fix this by requesting container data from the same node.
Asynchronous processing prevents deadlocks in this case. It also means that
failure to process a container is local now, not affecting other containers.

Signed-off-by: Roman Khimov <roman@nspcc.ru>