shard: collect expired tombstone objects separately from regular ones
Each expired _shard_ object currently triggers engine-level operations. These
are mostly relevant for split objects and are absolutely useless for
tombstones because tombstones have no children and they can't be locked.
Therefore TS-related operations can be local to shard, what we need to do is
to drop them, so mark them garbage for now and let the other GC part handle
it.

I believe this is one of the things that triggered #3507 since we have a lot of
deleted objects in some containers and iterating over TSes just takes time.

Signed-off-by: Roman Khimov <roman@nspcc.ru>