engine: buffer shard epoch events, allow to miss them
Fixes #3507 where an incomplete GC cycle blocks event handler which then breaks
WS client which then can't be restarted because we're still handling an event.
It's not a huge problem to miss this event if things are still moving, there
will be a new one with a new epoch and it'll do the job anyway.

Signed-off-by: Roman Khimov <roman@nspcc.ru>