fstree: fix buffer problem when extracting header
Due to the fact that the length of buffer did not correspond to the capacity and
did not increase anywhere, there were cases when there were fewer bytes in the
buffer than necessary. Fix problem. Add check for the length of slice in the
parsing of the combined prefix. Add test.

Closes #3517.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>