container: reduce the number of fetched iterator items
An attempt to list 2K+ containers ended up this way:

    error	handler/util.go:34	call method	{"status": 500, "request_id": "0cedec2b-935d-4575-a0c2-94ed4650c978", "method": "ListBuckets", "bucket": "", "object": "", "description": "something went wrong", "error": "list user containers via connection pool: status: code = 1024 message = json error: too big: size"}

We're trying to get 2K items from the iterator in a single call, each item is
base64 of CID, so it's 42 bytes, but then it's wrapped into VM type/value
struct like this:

    {"value":"","type":"ByteString"}

which adds another 32 bytes. For 2K elements this yields 144K JSON (not
including other data which is also present) which NeoGo refuses to return
given that the maximum size is 128K (https://github.com/nspcc-dev/neo-go/pull/3185).
Fetching in smaller chunks should fix the problem (for other iterator-based
methods as well).

Signed-off-by: Roman Khimov <roman@nspcc.ru>