storage/metabase: Check parent status in `ResolveECPart()`
It is the parent objects that are expiring and deleting in practice.
Reading is also performed at their address.

Previously, the method checked child status instead. In particular, this
led to 404 child status return when parent object was removed. This
collided with the part absence on SN: triggered requests from backup
SN and resulted in 404 final status.

Refs #3502.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>