gc: decouple expired object collector from epoch ticker
Move the processing of expired objects from the epoch event handler to the
regular GC cycle. The epoch event just sets some known epoch and marks that
expired objects need to be processed. When the GC runs, a certain number of
expired objects are immediately deleted (the limit is the same as for the GC).
If no more expired objects are found, the GC no longer checks for their presence
for the current epoch in subsequent ticks. Add GC remover sleep interval to the
tests to track removal of expired objects.

Closes #3522.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>