engine: fix flaky shard evacuate test
The test only monitors the last shard and stops when it reaches the target
count, but doesn't ensure uniform distribution across all shards. So, there is
situation when all 5 objects ended up in the 4th shard (the non-evacuated one),
while the first 3 shards (being evacuated) were empty. Since the engine check
guarantees that the last shard definitely has objects, we include it in the
evacuation list.

Closes #3516.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>