engine: store lock/tombstone object on every shard
Case: the lock is located on one shard and its information is recorded in the
metabucket of this shard, while the object being locked is located on another
shard and its information is recorded in the lock bucket. As a result, when
checking the second shard, the lock is located in the lock bucket, but it is not
possible to find out its expiration date, and it is reported that it is locked.
Therefore, store the lock object on all shards. The same situation may occur
with the tombstone object, so broadcast it too.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>