shard: exit early if there are no objects to delete
Regular GC cycle is very much noticeable in profiler even where there are no
objects to drop. The problem is that deleteObjs() is being called with zero
items and it opens a DB transaction even when it doesn't need one. DB
transaction has a substantial cost to it since it will perform some changes
to the file even if no key-values are changed. It can also affect other
threads since DB write transaction is exclusive.

Signed-off-by: Roman Khimov <roman@nspcc.ru>